+++ crates/coordintator-commons/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "coordinator-commons"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1"
+bdk = { version = "0.24.0" }
+serde = { version = "1", features = ["derive"] }
+time = { version = "0.3.20", features = ["serde"] }
+trade = { path = "../trade" }
+uuid = { version = "1.3.0", features = ["serde"] }
