+++ crates/trade/Cargo.toml
@@ -12,5 +12,3 @@ bdk = { version = "0.24.0" }
 rust_decimal = { version = "1", features = ["serde-with-float"] }
 rust_decimal_macros = "1.26"
 serde = { version = "1.0.152", features = ["serde_derive"] }
-time = { version = "0.3.20", features = ["serde"] }
-uuid = { version = "1.3.0", features = ["v4", "fast-rng", "macro-diagnostics", "serde"] }
